:truck: 컴포넌트에 정의 되어 있는 기능 함수 utils로 이동
prop-types 설치
