+++ front/package.json
@@ -11,6 +11,7 @@
     "antd-button-color": "^1.0.4",
     "dotenv": "^16.0.0",
     "moment": "^2.29.2",
+    "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-countdown-circle-timer": "^3.0.9",
     "react-dom": "^17.0.2",
